Add AccessControl class, enable ACL for ?quit cmd
Adds the AccessControl class, which does hostmask-based permissions checking
for running commands, and added AccessControl protection for the ?quit command.

Currently, the AccessControl list (or should I say HashMap?) is hard-coded to
grant all permissions to *!*@lizardirc/staff/* and nothing else.  This should
eventually be changed to pull in configuration a la java.util.Properties.

Change-Id: I5f13aaf5a7a3e1182b877c1a586b6b8acbe5650b