Add UserLastSeenListener
Provides the ?seen command (to ask the bot when it last saw someone talk in a
public channel) and the ?cfgseen command (to allow authorized users to
explicitly mark channels that shouldn't be tracked/considered "public" by the
bot).  The bot will also automatically not track channels set as secret
(channel mode +s).

Change-Id: Ieb71963c1eba5b4d19f2f4e490f3efef24babed8