Add reminders system
Adds the ?remind command, which allows you to remind someone else to do
something, and the ?remindme command, which allows you to remind yourself to do
something. Both commands support optional use of a time interval to have the
reminder be delivered at a specific point in the future, as opposed to the next
time you talk in or join a channel the bot is also in.

Also adds a ?clearreminders command, which does exactly what it says - clears
all waiting reminders out of the bot. Use with caution!

Change-Id: Ie3c6cb6db1e47c8acbc57f17057d2752d181251b