Allow SedListener to correct actions as well
This commit allows SedListener to correct channel actions in addition to channel
messages.  Note that as is, SedListener will still only accept replacements
given as a message, and replacements given as an action will be ignored (not
only will they not be acted on, but they won't be added to a user's message
history either).  This restriction is completely arbitrary though, and is easy
to remove if desired (see line 113).

Change-Id: I33918adf28a1a9222a49d7f33c640de617078cf1