Refactor commands into their own packages (2/2)
To ensure that file renames are logged properly, to avoid breaking file history,
this is changeset two in a sequence of two.  Both changesets *must* be submitted
together, in order, and this changeset must always be based on the first one.

This changeset contains the majority of the refactoring and renaming.  For the
three files that had to be manually separately renamed to preserve file history,
please see the first changeset of this set.

This commit refactors almost all Beancounter commands into their own packages
under the org.lizardirc.beancounter.commands package.  In addition, command
classes that formerly had private inner classes have now had those extracted
into package-local outer classes in their respective packages, and code has been
refactored as necessary to support this (this primarily consisted of changing
private methods in the existing outer classes to package-local methods as
needed, and adding package-local getters for some private fields of those
classes).

This should greatly improve the organization of the bot's source code as well as
code cleanliness, and has also vastly reduced line count bloat in the more
complex command handlers.

This commit also includes a few minor code cleanups here and there.

Resolves GitHub issue #7.

Change-Id: I7e9b9db7bda2b9cc8560ccdaa70c42b9bf4b1f86