Refactor commands into their own packages (1/2)
To ensure that file renames are logged properly, to avoid breaking file history,
this is changeset one in a sequence of two.  Both changesets *must* be submitted
together, in order, and the second changeset must always remain based on this
one.

This changeset contains only the required file renames and minimal changes to
ensure the build still succeedcs.  The remainder of the renames and refactoring
is in the second changeset of the set.

This commit refactors almost all Beancounter commands into their own packages
under the org.lizardirc.beancounter.commands package.  In addition, command
classes that formerly had private inner classes have now had those extracted
into package-local outer classes in their respective packages, and code has been
refactored as necessary to support this (this primarily consisted of changing
private methods in the existing outer classes to package-local methods as
needed, and adding package-local getters for some private fields of those
classes).

This should greatly improve the organization of the bot's source code as well as
code cleanliness, and has also vastly reduced line count bloat in the more
complex command handlers.

This commit also includes a few minor code cleanups here and there.

Resolves GitHub issue #7.

Change-Id: Ia743d4f3dedd42398d9fdab4451c903a868cd436