(WIP) Enable nickname substituion in ?lamborghini quotes
THIS COMMIT IS A WORK IN PROGRESS.  PLEASE DO NOT SUBMIT.

Adds support for substituting nicknames for chosen words in the ?lamborghini
command's Tai Lopez quotes.  The nickname selected is based on the command's
argument (similar to ?slap), and it falls back to a default if no argument is
provided.

Change-Id: I30ce8ffef70a6805db7b018197d8b3cd72ee3c08