Enable nickname substituion in ?lamborghini quotes
Adds support for substituting nicknames for chosen words in the ?lamborghini
command's Tai Lopez quotes.  The nickname selected is based on the command's
argument (similar to ?slap), and it falls back to a default if no argument is
provided.  If a replacement argument is provided, the bot also supports
applying one or more "transformers" to the replacement argument (e.g., make it
all-caps and append something to the end).

Change-Id: I30ce8ffef70a6805db7b018197d8b3cd72ee3c08