Update patch set 9
Patch Set 9: Code-Review+2 Verified+1

(1 comment)

Looks good and tested to work as expected.

One last thing you may wish to do which would be a potential improvement over HelixFossil's implementation is add a .trim() to the line noted below.  I consider this an optional improvement though; if you want to do it, great, but if not, just let me know and I'll submit as-is.

Patch-set: 9
Label: Verified=+1
Label: Code-Review=+2