Update patch set 4
Patch Set 4:

Nvm on that platform-dependent behavior, I'm just stupid.  TLUL correctly diagnosed it as the fact that without setting the proper config options, the ?weather command is never enabled, so ?w unambiguously resolved to ?wikipedia; but with the ?weather command enabled, ?w is ambiguous so the bot won't respond to it.  Derp.

Patch-set: 4