Add simplistic nickname ignore list for wikilink expansion
This is purposefully intended to be simplistic, as it's intention is not
one of security, rather convenience; apply this to IRC bots which
such as RC bots, as they already provide both wikilinks and URLs.

I'm also specifically not adapting this to json-based persistence,
when this is more concrete we can push this over as necessary.

Testing:
<stwalkerster> ?cfgwiki ignore stwalkerster
<PeaEnumerator> stwalkerster: Done
<stwalkerster> ?cfgwiki il
<PeaEnumerator> stwalkerster: Ignored users: stwalkerster
<stwalkerster> [[foo]]
<stwalkerster> ?cfgwiki ui stwalkerster
<PeaEnumerator> stwalkerster: Done
<stwalkerster> [[foo]]
<PeaEnumerator> stwalkerster: Foobar | The terms foobar (/ˈfuːbɑːr/), ...

Change-Id: If8c12fa71a35f87ee4f6a36f68cbfe16b7d38f68