(WIP) Add Elite: Dangerous commands
THIS IS A WORK IN PROGRESS.  DO NOT SUBMIT.

Adds the ?cmdr command to the bot to look up the information of an Elite:
Dangerous player from the Elite Dangerous Star Map (EDSM) - specifically, last
reported location, combat/merchant/exploration/CQC and naval ranks, and last
reported ship used.

Also adds the ability for the bot to stream FSDJump/Docked/Undocked journal
events from players submitting their data to the Elite Dangerous Data Network
(EDDN), and when added to a per-channel tracking list in the bot.

Change-Id: Iaa23c85f5f8a4a12297f0fc3f2a6fa515f6f8bcf