Fix morse decoder translating ambiguous characters
Previously, this would treat a message as containing morse if it
contained at least two morse characters. This includes things like
single dashes as punctuation.

Now, any ambiguous characters get stored in a buffer, and only
translated and applied to the message if there's another morse character
immediately following it, else the original input is applied without
modification.

I've also updated jUnit to move the tests for this to a data-driven
model instead of discrete test methods - it's easier to add test cases
and deduplicates what is essentially the exact same method with
different hardcoded values

Change-Id: Ib03a9c1d732473e2a353ac7d46b66b11f597ce28