Watch for a bot error condition that delivers reminders to everyone
Occasionally, due to a suspected bug in the upstream library's
(PircBotX) handling of connection failures, reminders that are
targeted at "" or "@" are created when the bot loses track of what
hostmasks users have.  In addition, if the bot is still having this
issue when the reminder is to be delivered, this reminder will then
match all users the bot can see on IRC (since all users will have the
hostmask "@"), who will then get the message delivered to them in PM.

This is, of course, somewhat suboptimal.

This change alters the matcher logic to detect this situation and
refuse to deliver the reminder, and alters the Reminder constructor to
also try to detect this situation and refuse to set reminders.  The
new constructor should catch most cases with a user-visible descriptive
error message, while the checks in the matcher should stop any cases
that slip through.

Change-Id: Icc587761cacbc106d118e0f22735f1edcd022003