Update patch set 4
Patch Set 4:

> There appears to be a regression introduced in this PS with the handling of word separators (/) between ambiguous morse symbols. For example, with Beancounter on master and EmbryonicPlantObserver built from this change:
> 
>  02:00:03 <@FastLizard4> ... / ... / - / ---
>  02:00:03 < Beancounter> FastLizard4 meant to say: S S T O
>  02:00:04 < EmbryonicPlantObserver> FastLizard4 meant to say: SS T O
> 
> Looks like this happens as long as the first character is an ambiguous character, regardless of the second character. Also happens if there's surrounding text.

So this was happening because it was caching the first S, then seeing a /, adding the space to the output, then seeing another morse char so adding the buffer to the output followed by the new character - ie: " SS T O". The fix is to treat the "/" as a buffer flush as well.

Patch-set: 4
Reviewer: Gerrit User 16 <16@cf5d0eaf-4814-484b-b55f-41cd7f22928e>