Update PircBotX and Google Client API SDK, with associated changes, to fix SedListener
Yak shaving: The lunchbox! Yak shaving: The breakfast cereal! Yak shaving, The *flamethrower!*

(The kids love this one!)

In seriousness, it turned out that SedListener was broken because we were including multiple incompatible versions of the Google Guava library, which turned out to be because PircBotX depends on (and still does, as of this commit, as 2.2 is the most recent stable release) a ridiculously old version of Guava, specifically 19.0. This breakage was introduced in commit f185c222 when we originally upgraded out Guava.

To fix this, we had to resolve the conflicting dependencies, which meant we had to upgrade PircBotX, unfortunately to the latest SNAPSHOT since apparently the bot hasn't had a normal release since late 2018. This in turn broke a lot of things, specifically because PircBotX removed all the generics from their various classes. While this is a good change, it also broke all the things which is annoying because it's just an upgrade within the 2.0 line but whatever.

Since I'm here, also did a bunch of small refactorings and cleanups here and there. Also updated a lot of the copyright headers.

Change-Id: I1f14d41b9dd505786882647fdfe9602d9f56b2a9